refactor name and locations for the EncryptorContainter
