+++ rpc/__tests__/transaction.ts
@@ -32,7 +32,7 @@ import { RPCTransaction } from '../transaction';
 import { RPCLoad } from '../Load/1.1.0/RPCLoad';
 import { Wallet } from "../../src/entity/Wallet";
 import { PrivateKeyDBFieldEncryption } from "../../src/entity/encryption/PrivateKeyDBFieldEncryption";
-import { EncryptorContainer } from '../../src/shipchain/EncryptorContainer';
+import { EncryptorContainer } from '../../src/entity/encryption/EncryptorContainer';
 
 export const RPCTransactions = async function() {
