+++ rpc/wallet.ts
@@ -19,8 +19,6 @@ import { BaseContract } from '../src/contracts/BaseContract';
 import { MetricsReporter } from '../src/MetricsReporter';
 import { LoadedContracts } from './contracts';
 import { RPCMethod, RPCNamespace } from './decorators';
-import { AwsPrivateKeyDBFieldEncryption } from '../src/shipchain/AwsPrivateKeyDBFieldEncryption';
-import { PrivateKeyDBFieldEncryption } from '../src/entity/encryption/PrivateKeyDBFieldEncryption';
 
 const loadedContracts = LoadedContracts.Instance;
 const metrics = MetricsReporter.Instance;
