+++ src/__tests__/contracts.ts
@@ -20,7 +20,7 @@ import 'mocha';
 import { Wallet } from '../entity/Wallet';
 import { Contract, Version, Project, Network } from '../entity/Contract';
 import { PrivateKeyDBFieldEncryption } from "../entity/encryption/PrivateKeyDBFieldEncryption";
-import { EncryptorContainer } from '../shipchain/EncryptorContainer';
+import { EncryptorContainer } from '../entity/encryption/EncryptorContainer';
 
 const utils = require('../local-test-net-utils');
