+++ src/shipchain/__tests__/loadvault.ts
@@ -21,7 +21,7 @@ import { LoadVault } from '../LoadVault';
 import { Wallet } from '../../entity/Wallet';
 import { PrivateKeyDBFieldEncryption } from "../../entity/encryption/PrivateKeyDBFieldEncryption";
 import { CloseConnection } from "../../redis";
-import { EncryptorContainer } from '../EncryptorContainer';
+import { EncryptorContainer } from '../../entity/encryption/EncryptorContainer';
 
 const storage_driver = { driver_type: 'local', base_path: 'storage/vault-tests' };
